+++ sentry/src/test/java/io/sentry/SentryTransactionTest.kt
@@ -167,4 +167,10 @@ class SentryTransactionTest {
         transaction.name = "new name"
         assertEquals("new name", transaction.transaction)
     }
+
+    @Test
+    fun `when transaction is not finished, status is null`() {
+        val transaction = SentryTransaction("name", "op")
+        assertNull(transaction.status)
+    }
 }